requirements.txt
Flask==2.3.2
openai==1.55.3
python-dotenv==1.0.0
